[Refactor] ProgramDetailVo  위치 이동
