+++ src/main/java/com/letsintern/letsintern/domain/program/helper/LetsChatHelper.java
@@ -3,7 +3,7 @@ package com.letsintern.letsintern.domain.program.helper;
 import com.letsintern.letsintern.domain.program.domain.LetsChat;
 import com.letsintern.letsintern.domain.program.exception.LetsChatNotFoundException;
 import com.letsintern.letsintern.domain.program.repository.LetsChatRepository;
-import com.letsintern.letsintern.domain.program.vo.ProgramDetailVo;
+import com.letsintern.letsintern.domain.program.vo.program.ProgramDetailVo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
